chore: Bump to MSBuild.Sdk.Extras 3.0.44
Context: https://github.com/novotnyllc/MSBuildSdkExtras/commit/d893038657560e2492fffb5e876fa9e5747345fd

Trying to build uno.toolkit.ui with .NET 10 would fail with
169 errors, e.g.:

	src/Uno.Toolkit.UI/Behaviors/SelectorExtensions.cs(142,3): error CS0436:
	  The type 'DynamicDependencyAttribute' in '…/uno.toolkit.ui/src/Uno.Toolkit.UI/Helpers/LinkerAttributes.cs' conflicts with the imported type 'DynamicDependencyAttribute' in 'System.Runtime, Version=9.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a'.
	  Using the type defined in '…/uno.toolkit.ui/src/Uno.Toolkit.UI/Helpers/LinkerAttributes.cs'.

The immediate cause of these errors was that when building
`Uno.Toolkit.WinUI.csproj`, `NETCOREAPP` was not defined.  This in
turn meant that all of the types within `LinkerAttributes.cs` were
now present in the net9.0 build, resulting in the CS0436 warnings.

`NETCOREAPP` was not defined, in turn, because
`$(DisableImplicitFrameworkDefines)` is `true`!

`$(DisableImplicitFrameworkDefines)` is true because the
MSBuild.Sdk.Extras nuget package set it!

Enter novotnyllc/MSBuildSdkExtras@d8930386, which *removes* the setting
of `$(DisableImplicitFrameworkDefines)`.  This change is present in
MSBuild.Sdk.Extras 3.0.44.

Bump to MSBuild.Sdk.Extras/3.0.44 so that
`$(DisableImplicitFrameworkDefines)` is no longer defined, allowing
uno.toolkit.ui to be built under .NET 10.